pyproject.toml
[project]
name = "compaction_analysis"
version = "0.1.0"
description = "Plots and simple models for RocksDB compaction analysis"
authors = [{name="Mobile-OS Lab", email=""}]
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.5",
    "numpy>=1.22",
    "pyyaml>=6.0"
]
